Working on cleaning up the AppInfo App's source code.
